Surveys blueprint refactoring
