+++ app/templates/surveys-access.html
@@ -35,7 +35,7 @@
 
 <form
         method="post"
-        action="{{ url_for('upload_survey', case_id=case_id, collection_instrument_id=collection_instrument_id, data=data) }}"
+        action="{{ url_for('surveys_bp.upload_survey', case_id=case_id, collection_instrument_id=collection_instrument_id, data=data) }}"
         class="form"
         role="form"
         enctype="multipart/form-data">
@@ -56,7 +56,7 @@
 <br/>
 
 <div>
-    <a href="/">Back</a>
+    <a href="{{ url_for('surveys_bp.logged_in') }}">Back</a>
 </div>
 
 {% endblock main %}
