+++ app/templates/surveys-history.html
@@ -8,11 +8,11 @@
 
 <ul class="navigation-tabs navigation-tabs-component">
     <li class="navigation-tabs__item navigation-tabs__item--primary">
-        <a href="/"
+        <a href="{{ url_for('surveys_bp.logged_in') }}"
             class="btn btn--secondary btn--border navigation-tabs__tab">To do</a>
     </li>
     <li class="navigation-tabs__item navigation-tabs__item--primary navigation-tabs__item--active">
-        <a href="/history"
+        <a href="{{ url_for('surveys_bp.surveys_history') }}"
             class="btn btn--secondary btn--border navigation-tabs__tab navigation-tabs__tab--active">History</a>
     </li>
     <li class="navigation-tabs__item navigation-tabs__item--primary">
