+++ app/templates/surveys-upload-success.html
@@ -14,8 +14,8 @@
             </div>
             <div class="panel__body">
                 <p>{{ file }} has been uploaded.</p>
-                <p>Business Register and Employment Survey has now moved into your <a href="/history">history</a>.</p>
-                <a href="/">Return to my to do list</a>
+                <p>Business Register and Employment Survey has now moved into your <a href="{{ url_for('surveys_bp.surveys_history') }}">history</a>.</p>
+                <a href="{{ url_for('surveys_bp.logged_in') }}">Return to my to do list</a>
             </div>
         </div>
     </div>
